Extract Project::FileLoader
