+++ lib/steep.rb
@@ -71,6 +71,7 @@ require "steep/project/file"
 require "steep/project/options"
 require "steep/project/target"
 require "steep/project/dsl"
+require "steep/project/file_loader"
 require "steep/drivers/utils/driver_helper"
 require "steep/drivers/check"
 require "steep/drivers/validate"
