+++ lib/steep/drivers/check.rb
@@ -20,8 +20,10 @@ module Steep
       def run
         project = load_config()
 
-        load_sources(project, command_line_patterns)
-        load_signatures(project)
+        loader = Project::FileLoader.new(project: project)
+        loader.load_sources(command_line_patterns)
+        loader.load_signatures()
+
         type_check(project)
 
         if self.dump_all_types
