+++ lib/steep/drivers/langserver.rb
@@ -23,8 +23,10 @@ module Steep
 
       def run
         project = load_config()
-        load_sources project, []
-        load_signatures project
+
+        loader = Project::FileLoader.new(project: project)
+        loader.load_sources([])
+        loader.load_signatures()
 
         reader.read do |request|
           Steep.logger.tagged "lsp" do
